added template for results. this is untested.
added results_date in contest schema for datetime that results are released. I
intend for this to be utcnow() if the admin chooses to release results
immediately.

All categories in the results.category column and all overlays in
results.category_overlay will be group together, and should be sorted
from highest to lowest score.

The sql call will need to change so we can join tables to get all
fields.